src/test/java/camp/nextstep/edu/rebellion/calc/StringAddCalculatorTest.java
package camp.nextstep.edu.rebellion.calc;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.junit.jupiter.api.Assertions.*;

class StringAddCalculatorTest {
    @DisplayName("빈문자 이거나 null일 경우 0을 반환")
    @Test
    public void splitAndSum_null_또는_빈문자() {
        int result = StringAddCalculator.splitAndSum(null);
        assertThat(result).isEqualTo(0);

        result = StringAddCalculator.splitAndSum("");
        assertThat(result).isEqualTo(0);
    }

    @DisplayName("숫자 하나 일경우 올바르게 연산되는지 확인")
    @Test
    public void splitAndSum_숫자하나() throws Exception {
        int result = StringAddCalculator.splitAndSum("1");
        assertThat(result).isEqualTo(1);
    }

    @DisplayName("쉼표 구분자 일 경우 올바르게 연산되는지 확인")
    @Test
    public void splitAndSum_쉼표구분자() throws Exception {
        int result = StringAddCalculator.splitAndSum("1,2");
        assertThat(result).isEqualTo(3);
    }

    @DisplayName("쉼표 구분자 혹은 콜론 구분자 일 경우 올바르게 연산되는지 확인")
    @Test
    public void splitAndSum_쉼표_또는_콜론_구분자() throws Exception {
        int result = StringAddCalculator.splitAndSum("1,2:3");
        assertThat(result).isEqualTo(6);
    }

    @DisplayName("커스텀 구분자 일 경우 올바르게 연산되는지 확인")
    @Test
    public void splitAndSum_custom_구분자() throws Exception {
        int result = StringAddCalculator.splitAndSum("//;\n1;2;3");
        assertThat(result).isEqualTo(6);
    }

    @DisplayName("음수일 경우 예외 발생")
    @Test
    public void splitAndSum_negative() throws Exception {
        assertThatThrownBy(() -> StringAddCalculator.splitAndSum("-1,2,3"))
                .isInstanceOf(RuntimeException.class)
                .hasMessage("음수 입니다");
    }

    @DisplayName("숫자형식이 아닌경우 예외 발생")
    @Test
    public void splitAndSum_notNumeric() {
        assertThatThrownBy(() -> StringAddCalculator.splitAndSum("A,2,3"))
                .isInstanceOf(RuntimeException.class)
                .hasMessageContaining("For input string");
    }
}
